516: queue-box WIP
