+++ web-client/src/styles/tabs.scss
@@ -1,15 +1,26 @@
-.work-queue-tab-container {
+.queue-tab-container {
   border: 1px solid $color-gray-lighter;
   background-color: $color-gray-lightest;
-}
 
-.work-queue-tab {
-  display: inline-block;
-  padding: 1rem 2rem;
-  border-right: 1px solid $color-gray-lighter;
-  margin: 0;
-  background-color: $color-white;
-  box-shadow: 0 2px 0 $color-white;
+  .queue-tab {
+    box-sizing: border-box;
+    font-family: $font-serif;
+    font-size: 20px;
+    font-weight: 500;
+    line-height: 1.3em;
+    display: inline-block;
+    padding: 1rem 2rem;
+    margin: 0;
+    border-radius: 0;
+
+    &.selected,
+    &[aria-selected='true'] {
+      font-weight: 700;
+      border-right: 1px solid $color-gray-lighter;
+      background-color: $color-white;
+      box-shadow: 0 2px 0 $color-white;
+    }
+  }
 }
 
 .horizontal-tabs {
