+++ web-client/src/views/WorkQueue.jsx
@@ -24,10 +24,12 @@ export default connect(
                 role="tab"
                 className="tab-link"
                 id="tab-my-queue"
+                aria-controls="tab-individual-panel"
                 aria-selected={workQueueHelper.showIndividualWorkQueue}
                 onClick={() =>
                   switchWorkQueueSequence({
-                    workQueueToDisplay: 'individual',
+                    queue: 'my',
+                    box: 'inbox',
                   })
                 }
               >
@@ -41,10 +43,12 @@ export default connect(
                 role="tab"
                 className="tab-link"
                 id="tab-work-queue"
+                aria-controls="tab-section-panel"
                 aria-selected={workQueueHelper.showSectionWorkQueue}
                 onClick={() =>
                   switchWorkQueueSequence({
-                    workQueueToDisplay: 'section',
+                    queue: 'section',
+                    box: 'inbox',
                   })
                 }
               >
@@ -53,17 +57,15 @@ export default connect(
             </li>
           </ul>
         </nav>
-        <div className="work-queue-tab-container">
-          <h3 className="work-queue-tab">Inbox</h3>
-        </div>
-        {workQueueHelper.showSectionWorkQueue && (
-          <div role="tabpanel">
-            <SectionWorkQueue />
+
+        {workQueueHelper.showIndividualWorkQueue && (
+          <div role="tabpanel" id="tab-individual-panel">
+            <IndividualWorkQueue />
           </div>
         )}
-        {workQueueHelper.showIndividualWorkQueue && (
-          <div role="tabpanel">
-            <IndividualWorkQueue />
+        {workQueueHelper.showSectionWorkQueue && (
+          <div role="tabpanel" id="tab-section-panel">
+            <SectionWorkQueue />
           </div>
         )}
       </React.Fragment>
